requirements/requirements.txt
aiobotocore==2.21.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beanie==1.29.0
botocore==1.37.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
docutils==0.16
fastapi==0.115.12
frozenlist==1.6.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
iniconfig==2.1.0
isodate==0.6.1
jmespath==1.0.1
lazy-model==0.2.0
Mako==1.3.10
MarkupSafe==3.0.2
motor==3.7.0
multidict==6.4.3
packaging==25.0
pluggy==1.5.0
propcache==0.3.1
pyasn1==0.6.1
pydantic==1.10.22
pydantic_core==2.33.1
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rfc3986==1.5.0
rsa==4.7.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
toml==0.10.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wrapt==1.17.2
yarl==1.20.0
